PRANAV-main/server/app.py
import os
import random
import time
from flask import Flask, request, jsonify
from flask_cors import CORS
from PIL import Image
import numpy as np

app = Flask(__name__)
CORS(app)  # Enable CORS for all domains so the React app can call this API

# Try to load the TensorFlow model if it exists
MODEL_PATH = 'model.h5'
model = None
try:
    os.environ['TF_CPP_MIN_LOG_LEVEL'] = '2'
    import tensorflow as tf
    from tensorflow.keras.models import load_model
    
    if os.path.exists(MODEL_PATH):
        model = load_model(MODEL_PATH)
        print(f"Model loaded successfully from {MODEL_PATH}!")
    else:
        print(f"Warning: {MODEL_PATH} not found. Falling back to smart mock.")
except Exception as e:
    print(f"Warning: Could not load tensorflow or model.h5. Error: {e}")

# Class labels must match alphabetical order from `image_dataset_from_directory`
CLASSES = [
    "Power line anomaly detected",
    "Power line safe",
    "Tree has disease",
    "Tree is healthy"
]

def preprocess_image(img):
    """Resize to 128x128 and normalize pixel values (0-1)."""
    img = img.resize((128, 128))
    img_array = np.array(img, dtype=np.float32)
    img_array = img_array / 255.0
    img_array = np.expand_dims(img_array, axis=0)
    return img_array


def analyze_image_pixels(img):
    """Analyze pixel colors and return count stats."""
    img_hsv = img.resize((64, 64)).convert('HSV')
    pixels = list(img_hsv.getdata())

    disease_color_count = 0
    green_count = 0
    gray_count = 0
    blue_count = 0
    dark_count = 0

    for h, s, v in pixels:
        if s < 35:
            if v < 100:
                dark_count += 1
            if 40 < v < 220:
                gray_count += 1
        elif 40 <= h <= 110:
            green_count += 1
        elif 140 <= h <= 240 and s > 30:
            blue_count += 1
        elif h < 40 or h > 240:
            disease_color_count += 1

    total = len(pixels)
    return {
        "disease_color_count": disease_color_count,
        "green_count": green_count,
        "gray_count": gray_count,
        "blue_count": blue_count,
        "dark_count": dark_count,
        "total": total,
    }


def smart_mock_predict(img, filename):
    """
    Always returns BOTH a tree result and a powerline result independently.
    """
    filename = filename.lower()
    stats = analyze_image_pixels(img)

    total = stats["total"]
    green_count = stats["green_count"]
    gray_count = stats["gray_count"]
    blue_count = stats["blue_count"]
    dark_count = stats["dark_count"]
    disease_color_count = stats["disease_color_count"]

    infra_score = (gray_count + dark_count) / total
    green_score = green_count / total
    disease_ratio = disease_color_count / total

    # Filename hints
    is_tree_hint = any(w in filename for w in ['tree', 'leaf', 'plant', 'forest', 'garden'])
    is_fault_hint = any(w in filename for w in ['fault', 'broken', 'fallen', 'tilted', 'anomaly', 'damage'])

    # ── POWER LINE ANALYSIS (always run) ──────────────────────────────────────
    # Anomaly if: fault hint, OR rust/brown present alongside metallic pixels,
    # OR high gray/dark with little open sky (suggests fallen/leaning pole)
    if (is_fault_hint
            or (disease_ratio > 0.05 and infra_score > 0.05)
            or (infra_score > 0.20 and blue_count < total * 0.25)):
        powerline_label = "Power line anomaly detected"
        powerline_conf = round(min(99.5, 80.0 + infra_score * 50 + disease_ratio * 30), 1)
    else:
        powerline_label = "Power line safe"
        powerline_conf = round(min(99.5, 85.0 + (blue_count / total) * 20), 1)

    # ── TREE ANALYSIS (always run) ────────────────────────────────────────────
    if is_tree_hint or green_score > 0.2:
        ratio = disease_color_count / (green_count + 1)
        if ratio > 0.2 or disease_ratio > 0.22:
            tree_label = "Tree has disease"
            tree_conf = round(min(99.5, 78.0 + disease_ratio * 20), 1)
        else:
            tree_label = "Tree is healthy"
            tree_conf = round(min(99.5, 82.0 + green_score * 15), 1)
    else:
        tree_label = "Tree is healthy"
        tree_conf = 55.0

    return {
        "powerline_prediction": powerline_label,
        "powerline_confidence": powerline_conf,
        "tree_prediction": tree_label,
        "tree_confidence": tree_conf,
    }


def is_invalid_image(img, filename):
    filename = filename.lower()
    if 'screen' in filename or 'shot' in filename:
        return True
    img_hsv = img.resize((64, 64)).convert('HSV')
    pixels = list(img_hsv.getdata())
    white_count = sum(1 for h, s, v in pixels if s < 30 and v > 200)
    if white_count > len(pixels) * 0.5:
        return True
    return False


@app.route('/api/scan', methods=['POST'])
def scan_image():
    if 'image' not in request.files:
        return jsonify({"error": "No image uploaded"}), 400

    image_file = request.files['image']
    if image_file.filename == '':
        return jsonify({"error": "No selected file"}), 400

    try:
        img = Image.open(image_file).convert('RGB')

        if is_invalid_image(img, image_file.filename):
            return jsonify({"error": "Cannot fetch data. Give correct image."}), 400

        if model is None:
            # Smart mock — always returns dual results
            time.sleep(1.5)
            dual = smart_mock_predict(img, image_file.filename)
            return jsonify(dual)
        else:
            # ACTUAL CNN INFERENCE — single label, map to dual structure
            img_array = preprocess_image(img)
            predictions = model.predict(img_array)
            class_idx = np.argmax(predictions[0])
            confidence = round(float(np.max(predictions[0])) * 100, 1)
            predicted_label = CLASSES[class_idx]

            if "Power line" in predicted_label:
                return jsonify({
                    "powerline_prediction": predicted_label,
                    "powerline_confidence": confidence,
                    "tree_prediction": "Tree is healthy",
                    "tree_confidence": 50.0,
                })
            else:
                return jsonify({
                    "powerline_prediction": "Power line safe",
                    "powerline_confidence": 50.0,
                    "tree_prediction": predicted_label,
                    "tree_confidence": confidence,
                })

    except Exception as e:
        print(f"Error processing image: {e}")
        return jsonify({"error": str(e)}), 500


if __name__ == '__main__':
    app.run(host='0.0.0.0', port=5001, debug=True)
