PRANAV-main/server/requirements.txt
Flask
Flask-Cors
Pillow
numpy
tensorflow
matplotlib
scikit-learn
gunicorn
